poc/pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepgram-sdk>=5.3.2",
    "deepl>=1.27.0",
    "python-dotenv>=1.2.1",
    "websockets>=16.0",
]
